Get a single localized value from the dates dataset

The dates table contains information on how to express dates and this data is localized across 574 locales. The cldr_dates() function allows one to extract a named list using a locale and a specific element. The element values are:

"months_format_abbrev"

"months_format_narrow"

"months_format_wide"

"days_standalone_narrow"

"days_standalone_short"

"days_standalone_wide"

"quarters_format_abbrev"

"quarters_format_narrow"

"quarters_format_wide"

"quarters_standalone_abbrev"

"quarters_standalone_narrow"

"quarters_standalone_wide"

"dayperiods_format_abbrev"

"dayperiods_format_narrow"

"dayperiods_format_wide"

"dayperiods_standalone_abbrev"

"dayperiods_standalone_narrow"

"dayperiods_standalone_wide"

"eras_abbrev"

"eras_names"

"eras_narrow"

"date_formats"

"date_skeletons"

"time_formats"

"time_skeletons"

"date_time_available_formats"

"date_time_append_items"

"date_time_interval_formats"

Usage

cldr_dates(locale = "en", element = dates_elements$months_format_abbrev)

Arguments

locale

The locale ID for which to obtain the data from the dates table.

element

The element from which information will be obtained for the specified locale.

Value

A named list.